backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai
pytesseract==0.3.10
Pillow>=10.0.0
gspread==5.12.0
oauth2client==4.1.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0